requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data validation and serialization
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Environment variables
python-dotenv==1.0.0

# HTTP client
httpx==0.25.2

# Date and time utilities
python-dateutil==2.8.2

# UUID utilities
uuid==1.30

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code formatting and linting
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Monitoring and logging
structlog==23.2.0

# CORS middleware
fastapi-cors==0.0.6

# Background tasks
celery==5.3.4
redis==5.0.1

# File handling
python-magic==0.4.27

# Email support
fastapi-mail==1.4.1

# Rate limiting
slowapi==0.1.9

# Health checks
psutil==5.9.6